Actually writing classifications to output vector.
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rulebasedmodels@223 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ C5/R/predict.C5.0.R
@@ -20,7 +20,7 @@ predict.C5.0 <- function (object, newdata = NULL, ...)
           as.character(object$names),
           as.character(object$tree),
           as.character(object$rules),
-          pred = double(nrow(newdata)),  # XXX predictions are character
+          pred = integer(nrow(newdata)),  # XXX should this be character?
           output = character(1),
           PACKAGE = "C50"
           )
